test(api): split the 2,059-line api.test.ts into per-domain files
Mirror the route modules: one test file per domain over a shared test/helpers.ts.
Pure move — every assertion and test name is verbatim (127 it() before = 127
after; 135 api tests total, unchanged).

- test/helpers.ts: the shared harness (open-mode app/meta/dir + afterAll cleanup,
  upload/postJson/json) and the auth/multi-tenant factories (makeAuthedApp, as,
  publishAs, jsonAs, proposeAs, bearer, quotaApp, pub, TestUser). vitest isolates
  each file, so every test file gets its own store + cleanup.
- artifacts / comments / reviews / analytics / permissions / agents / moderation /
  security / collections / favorites / workspace / quotas .test.ts — block-local
  helpers stay with their block.
- webhooks.test.ts + bus.test.ts untouched.

Gate green: typecheck, biome 0 errors, 14 files / 135 tests pass.
