--- apps/api/test/api.test.ts
@@ -1,2059 +1,0 @@
-import { mkdtempSync, rmSync } from "node:fs"
-import { tmpdir } from "node:os"
-import { join } from "node:path"
-import { SqliteMetaStore } from "@dock/db/sqlite"
-import { FsBlobStore } from "@dock/storage/fs"
-import Database from "better-sqlite3"
-import { zipSync } from "fflate"
-import { afterAll, describe, expect, it } from "vitest"
-import { type AppDeps, createApp, isPublicHttpUrl } from "../src/app"
-
-const dir = mkdtempSync(join(tmpdir(), "dock-test-"))
-const meta = new SqliteMetaStore(join(dir, "dock.db"))
-const app = createApp({
-  meta,
-  blobs: new FsBlobStore(join(dir, "blobs")),
-  baseUrl: "http://dock.test",
-})
-
-afterAll(() => {
-  meta.close()
-  rmSync(dir, { recursive: true, force: true })
-})
-
-const upload = (
-  name: string,
-  content: Uint8Array | string,
-  fields: Record<string, string> = {},
-  shortId?: string,
-) => {
-  const form = new FormData()
-  const bytes = typeof content === "string" ? new TextEncoder().encode(content) : content
-  form.append("file", new Blob([bytes as BlobPart]), name)
-  for (const [k, v] of Object.entries(fields)) form.append(k, v)
-  const url = shortId ? `/v1/artifacts/${shortId}/versions` : "/v1/artifacts"
-  return app.request(url, { method: "POST", body: form })
-}
-
-const postJson = (path: string, body: unknown) =>
-  app.request(path, {
-    method: "POST",
-    headers: { "content-type": "application/json" },
-    body: JSON.stringify(body),
-  })
-
-describe("version sessions", () => {
-  it("a named publish stores the checkpoint name on the version", async () => {
-    const { short_id } = await (await upload("n.md", "v1")).json()
-    await upload("n.md", "v2", { name: "Final draft" }, short_id)
-    const a = await (await app.request(`/v1/artifacts/${short_id}`)).json()
-    expect(a.versions[1].name).toBe("Final draft")
-    expect(a.versions[0].name).toBeNull()
-  })
-
-  it("includes a sessions array alongside raw versions", async () => {
-    const { short_id } = await (await upload("s.md", "v1")).json()
-    const a = await (await app.request(`/v1/artifacts/${short_id}`)).json()
-    expect(Array.isArray(a.sessions)).toBe(true)
-    expect(a.sessions[0]).toMatchObject({ n: 1, from_n: 1, count: 1 })
-  })
-
-  it("collapses a same-author burst into one session", async () => {
-    const { short_id } = await (await upload("b.md", "v1")).json()
-    await upload("b.md", "v2", {}, short_id)
-    await upload("b.md", "v3", {}, short_id)
-    const a = await (await app.request(`/v1/artifacts/${short_id}`)).json()
-    expect(a.versions).toHaveLength(3)
-    expect(a.sessions).toHaveLength(1)
-    expect(a.sessions[0]).toMatchObject({ n: 3, from_n: 1, count: 3 })
-  })
-
-  it("pins a named checkpoint as its own session", async () => {
-    const { short_id } = await (await upload("p.md", "v1")).json()
-    await upload("p.md", "v2", {}, short_id)
-    await upload("p.md", "v3", { name: "Approved" }, short_id)
-    const a = await (await app.request(`/v1/artifacts/${short_id}`)).json()
-    // newest-first: [named v3], [v1+v2 burst]
-    expect(a.sessions).toHaveLength(2)
-    expect(a.sessions[0]).toMatchObject({ n: 3, name: "Approved", count: 1 })
-    expect(a.sessions[1]).toMatchObject({ n: 2, from_n: 1, count: 2 })
-  })
-
-  it("starts a new session when the author changes", async () => {
-    const { short_id } = await (await upload("au.md", "v1", { author: "ava" })).json()
-    await upload("au.md", "v2", { author: "bo" }, short_id)
-    const a = await (await app.request(`/v1/artifacts/${short_id}`)).json()
-    expect(a.sessions).toHaveLength(2)
-    expect(a.sessions[0]).toMatchObject({ author: "bo" })
-    expect(a.sessions[1]).toMatchObject({ author: "ava" })
-  })
-
-  it("honors the configured window (each revision its own session)", async () => {
-    const m2 = new SqliteMetaStore(join(dir, "win.db"))
-    const app2 = createApp({
-      meta: m2,
-      blobs: new FsBlobStore(join(dir, "blobs")),
-      baseUrl: "http://dock.test",
-      versionWindowMs: -1,
-    })
-    const mk = (c: string, id?: string) => {
-      const fd = new FormData()
-      fd.append("file", new Blob([new TextEncoder().encode(c)]), "w.md")
-      return app2.request(id ? `/v1/artifacts/${id}/versions` : "/v1/artifacts", {
-        method: "POST",
-        body: fd,
-      })
-    }
-    const { short_id } = await (await mk("v1")).json()
-    await mk("v2", short_id)
-    const a = await (await app2.request(`/v1/artifacts/${short_id}`)).json()
-    expect(a.versions).toHaveLength(2)
-    expect(a.sessions).toHaveLength(2) // window -1 → never merge
-    m2.close()
-  })
-})
-
-describe("version restore", () => {
-  it("restores a past version as a new current revision", async () => {
-    const { short_id } = await (await upload("r.md", "alpha")).json()
-    await upload("r.md", "beta", {}, short_id)
-    const res = await postJson(`/v1/artifacts/${short_id}/restore`, { version: 1 })
-    expect(res.status).toBe(201)
-    const a = await res.json()
-    expect(a.current_version).toBe(3)
-    expect(a.versions[2].message).toBe("Restored v1")
-  })
-
-  it("reproduces the restored version's content exactly", async () => {
-    const { short_id } = await (await upload("rc.md", "# original")).json()
-    await upload("rc.md", "# changed", {}, short_id)
-    await postJson(`/v1/artifacts/${short_id}/restore`, { version: 1 })
-    const current = await (await app.request(`/v1/artifacts/${short_id}/content`)).text()
-    expect(current).toBe("# original")
-  })
-
-  it("preserves the original version after restore (history not rewritten)", async () => {
-    const { short_id } = await (await upload("rp.md", "one")).json()
-    await upload("rp.md", "two", {}, short_id)
-    await postJson(`/v1/artifacts/${short_id}/restore`, { version: 1 })
-    expect(await (await app.request(`/v1/artifacts/${short_id}/content?v=1`)).text()).toBe("one")
-    expect(await (await app.request(`/v1/artifacts/${short_id}/content?v=2`)).text()).toBe("two")
-  })
-
-  it("404s restoring an unknown version", async () => {
-    const { short_id } = await (await upload("r4.md", "x")).json()
-    expect((await postJson(`/v1/artifacts/${short_id}/restore`, { version: 99 })).status).toBe(404)
-  })
-
-  it("400s when no version is given", async () => {
-    const { short_id } = await (await upload("r0.md", "x")).json()
-    expect((await postJson(`/v1/artifacts/${short_id}/restore`, {})).status).toBe(400)
-  })
-})
-
-describe("publish html file", () => {
-  let shortId: string
-
-  it("publishes and returns a stable url", async () => {
-    const res = await upload(
-      "q1-review.html",
-      "<h1>Q1 Review</h1><script>document.title='hi'</script>",
-      {
-        title: "Q1 Review",
-      },
-    )
-    expect(res.status).toBe(201)
-    const json = await res.json()
-    shortId = json.short_id
-    expect(json.url).toBe(`http://dock.test/a/${shortId}-q1-review`)
-    expect(json.kind).toBe("file")
-    expect(json.current_version).toBe(1)
-  })
-
-  it("serves the viewer shell and sandboxed raw content", async () => {
-    const viewer = await app.request(`/a/${shortId}-q1-review`)
-    expect(viewer.status).toBe(200)
-    expect(await viewer.text()).toContain("Q1 Review")
-
-    const raw = await app.request(`/raw/${shortId}/v/1/index.html`)
-    expect(raw.status).toBe(200)
-    expect(raw.headers.get("content-security-policy")).toContain("sandbox allow-scripts")
-    expect(raw.headers.get("content-security-policy")).not.toContain("allow-same-origin")
-    expect(await raw.text()).toContain("<h1>Q1 Review</h1>")
-  })
-
-  it("republishes as v2 while @v1 stays immutable", async () => {
-    const res = await upload(
-      "q1-review.html",
-      "<h1>Q1 Review v2</h1>",
-      { message: "address review" },
-      shortId,
-    )
-    expect(res.status).toBe(201)
-    expect((await res.json()).current_version).toBe(2)
-
-    expect(await (await app.request(`/raw/${shortId}/v/1/index.html`)).text()).toContain(
-      "Q1 Review</h1>",
-    )
-    expect(await (await app.request(`/raw/${shortId}/v/2/index.html`)).text()).toContain(
-      "Q1 Review v2",
-    )
-
-    const oldViewer = await app.request(`/a/${shortId}@v1`)
-    expect(await oldViewer.text()).toContain("jump to current (v2)")
-  })
-})
-
-describe("publish static bundle (astro-style dist)", () => {
-  let shortId: string
-
-  it("publishes a zip with nested assets and pretty urls", async () => {
-    const zip = zipSync({
-      "index.html": new TextEncoder().encode("<h1>Site</h1><script src='/assets/app.js'></script>"),
-      "assets/app.js": new TextEncoder().encode("console.log('hi')"),
-      "about/index.html": new TextEncoder().encode("<h1>About</h1>"),
-    })
-    const res = await upload("dist.zip", zip, { title: "My Site", spa: "true" })
-    expect(res.status).toBe(201)
-    const json = await res.json()
-    shortId = json.short_id
-    expect(json.kind).toBe("bundle")
-  })
-
-  it("serves nested assets with correct mime", async () => {
-    const js = await app.request(`/raw/${shortId}/v/1/assets/app.js`)
-    expect(js.status).toBe(200)
-    expect(js.headers.get("content-type")).toContain("text/javascript")
-  })
-
-  it("rewrites root-absolute urls so assets resolve under the raw prefix", async () => {
-    const html = await (await app.request(`/raw/${shortId}/v/1/index.html`)).text()
-    expect(html).toContain(`src='/raw/${shortId}/v/1/assets/app.js'`)
-    expect(html).not.toContain(`src='/assets/app.js'`)
-  })
-
-  it("supports pretty urls and spa fallback", async () => {
-    const about = await app.request(`/raw/${shortId}/v/1/about`)
-    expect(await about.text()).toContain("<h1>About</h1>")
-
-    const fallback = await app.request(`/raw/${shortId}/v/1/some/client/route`)
-    expect(fallback.status).toBe(200)
-    expect(await fallback.text()).toContain("<h1>Site</h1>")
-  })
-})
-
-describe("view analytics", () => {
-  it("de-dups rapid re-opens and aggregates uniques, per-version, recent viewers", async () => {
-    const { short_id } = await (await upload("a.md", "# A")).json()
-    await upload("a.md", "# A v2", { message: "v2" }, short_id) // now at v2
-
-    // Three rapid opens from one anonymous viewer (cookie reused) collapse to ONE
-    // recorded view — a refresh no longer inflates the count.
-    let cookie = ""
-    for (let i = 0; i < 3; i++) {
-      const r = await app.request(`/v1/artifacts/${short_id}/view`, {
-        method: "POST",
-        headers: { "content-type": "application/json", ...(cookie ? { cookie } : {}) },
-        body: JSON.stringify({ version: 1 }),
-      })
-      cookie ||= (r.headers.get("set-cookie") ?? "").split(";")[0]
-    }
-    // The same viewer on a different version is a distinct view.
-    await app.request(`/v1/artifacts/${short_id}/view`, {
-      method: "POST",
-      headers: { "content-type": "application/json", cookie },
-      body: JSON.stringify({ version: 2 }),
-    })
-    // A fresh anonymous viewer (no cookie) is a distinct unique.
-    await app.request(`/v1/artifacts/${short_id}/view`, {
-      method: "POST",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({ version: 2 }),
-    })
-
-    const a = await (await app.request(`/v1/artifacts/${short_id}/analytics`)).json()
-    expect(a.total).toBe(3) // viewerA@v1 (de-duped from 3), viewerA@v2, viewerB@v2
-    expect(a.unique).toBe(2) // two distinct anon cookies
-    expect(a.perVersion).toEqual([
-      { version: 1, count: 1 },
-      { version: 2, count: 2 },
-    ])
-    expect(a.daily.reduce((s: number, d: { count: number }) => s + d.count, 0)).toBe(3)
-    expect(a.recent.length).toBe(2) // per-viewer, newest-first
-    expect(a.recent.every((r: { kind: string }) => r.kind === "anon")).toBe(true)
-
-    // Batch counts surface on the library listing.
-    const list = await (await app.request("/v1/artifacts")).json()
-    const row = list.artifacts.find((x: { short_id: string }) => x.short_id === short_id)
-    expect(row.views).toBe(3)
-  })
-
-  it("no-ops when analytics is disabled", async () => {
-    const off = createApp({
-      meta,
-      blobs: new FsBlobStore(join(dir, "blobs2")),
-      baseUrl: "http://dock.test",
-      analytics: false,
-    })
-    const { short_id } = await (async () => {
-      const form = new FormData()
-      form.append("file", new Blob([new TextEncoder().encode("# B")]), "b.md")
-      return (await off.request("/v1/artifacts", { method: "POST", body: form })).json()
-    })()
-    const v = await off.request(`/v1/artifacts/${short_id}/view`, {
-      method: "POST",
-      headers: { "content-type": "application/json" },
-      body: "{}",
-    })
-    expect(v.status).toBe(204)
-    const a = await off.request(`/v1/artifacts/${short_id}/analytics`)
-    expect(a.status).toBe(404)
-  })
-})
-
-describe("publish markdown", () => {
-  it("renders sanitized html and serves raw source", async () => {
-    const md = "# Notes\n\nSome *text*.\n\n<script>alert(1)</script>"
-    const res = await upload("notes.md", md)
-    const { short_id } = await res.json()
-
-    const rendered = await app.request(`/raw/${short_id}/v/1/index.html`)
-    const html = await rendered.text()
-    expect(html).toContain("<h1>Notes</h1>")
-    expect(html).not.toContain("<script>alert")
-
-    const raw = await app.request(`/raw/${short_id}/v/1/raw.md`)
-    expect(raw.headers.get("content-type")).toContain("text/markdown")
-    expect(await raw.text()).toBe(md)
-  })
-})
-
-describe("api surface", () => {
-  it("returns artifact json with version history", async () => {
-    const res = await upload("doc.html", "<p>one</p>", { title: "Doc" })
-    const { short_id } = await res.json()
-    await upload("doc.html", "<p>two</p>", { message: "tweak" }, short_id)
-
-    const meta = await (await app.request(`/v1/artifacts/${short_id}`)).json()
-    expect(meta.current_version).toBe(2)
-    expect(meta.versions).toHaveLength(2)
-    expect(meta.versions[1].message).toBe("tweak")
-  })
-
-  it("reads back source content for any version", async () => {
-    const res = await upload("read.md", "# one", { title: "Read" })
-    const { short_id } = await res.json()
-    await upload("read.md", "# two", { message: "v2" }, short_id)
-
-    const cur = await app.request(`/v1/artifacts/${short_id}/content`)
-    expect(cur.status).toBe(200)
-    expect(cur.headers.get("x-dock-version")).toBe("2")
-    expect(await cur.text()).toBe("# two")
-
-    const v1 = await app.request(`/v1/artifacts/${short_id}/content?v=1`)
-    expect(await v1.text()).toBe("# one")
-  })
-
-  it("reads back a bundle's entry document", async () => {
-    const zip = zipSync({ "index.html": new TextEncoder().encode("<h1>Entry</h1>") })
-    const { short_id } = await (await upload("site.zip", zip)).json()
-    const content = await app.request(`/v1/artifacts/${short_id}/content`)
-    expect(content.headers.get("x-dock-kind")).toBe("bundle")
-    expect(await content.text()).toBe("<h1>Entry</h1>")
-  })
-
-  it("diffs two versions as text and json", async () => {
-    const res = await upload("d.md", "# title\nalpha", { title: "D" })
-    const { short_id } = await res.json()
-    await upload("d.md", "# title\nbeta", { message: "v2" }, short_id)
-
-    const txt = await app.request(`/v1/artifacts/${short_id}/diff`)
-    expect(txt.status).toBe(200)
-    expect(txt.headers.get("x-dock-from")).toBe("1")
-    expect(txt.headers.get("x-dock-to")).toBe("2")
-    const body = await txt.text()
-    expect(body).toContain("  # title")
-    expect(body).toContain("- alpha")
-    expect(body).toContain("+ beta")
-
-    const json = await (await app.request(`/v1/artifacts/${short_id}/diff?format=json`)).json()
-    expect(json.from).toBe(1)
-    expect(json.to).toBe(2)
-    expect(json.ops).toContainEqual({ t: "add", line: "beta" })
-  })
-
-  it("404s on unknown artifacts and rejects empty zips", async () => {
-    expect((await app.request("/v1/artifacts/zzzzzzzz")).status).toBe(404)
-    expect((await app.request("/a/zzzzzzzz")).status).toBe(404)
-    const bad = await upload("dist.zip", zipSync({}))
-    expect(bad.status).toBe(400)
-  })
-})
-
-const json = (obj: unknown) => ({
-  method: "POST" as const,
-  headers: { "content-type": "application/json" },
-  body: JSON.stringify(obj),
-})
-
-describe("comments + the loop", () => {
-  let shortId: string
-  let threadId: string
-  let rootId: string
-
-  it("creates a comment as a new thread", async () => {
-    shortId = (await (await upload("c.md", "# doc with mean sentiment", { title: "C" })).json())
-      .short_id
-    const res = await app.request(
-      `/v1/artifacts/${shortId}/comments`,
-      json({
-        body_md: "use median",
-        author: "jess",
-        anchor: { type: "TextQuoteSelector", exact: "mean" },
-      }),
-    )
-    expect(res.status).toBe(201)
-    const cm = await res.json()
-    expect(cm.state).toBe("open")
-    expect(cm.thread_id).toBe(cm.id)
-    expect(cm.base_version).toBe(1)
-    expect(cm.anchor).toContain("TextQuoteSelector")
-    threadId = cm.thread_id
-    rootId = cm.id
-  })
-
-  it("replies in the same thread", async () => {
-    const cm = await (
-      await app.request(
-        `/v1/artifacts/${shortId}/comments`,
-        json({ body_md: "agreed", thread_id: threadId }),
-      )
-    ).json()
-    expect(cm.thread_id).toBe(threadId)
-    const list = await (await app.request(`/v1/artifacts/${shortId}/comments`)).json()
-    expect(list.comments).toHaveLength(2)
-  })
-
-  it("resolves and reopens a thread, with state filtering", async () => {
-    await app.request(`/v1/artifacts/${shortId}/comments/${rootId}/resolve`, { method: "POST" })
-    expect(
-      (await (await app.request(`/v1/artifacts/${shortId}/comments?state=open`)).json()).comments,
-    ).toHaveLength(0)
-    expect(
-      (await (await app.request(`/v1/artifacts/${shortId}/comments?state=resolved`)).json())
-        .comments,
-    ).toHaveLength(2)
-
-    await app.request(
-      `/v1/artifacts/${shortId}/comments/${rootId}/resolve`,
-      json({ state: "open" }),
-    )
-    expect(
-      (await (await app.request(`/v1/artifacts/${shortId}/comments?state=open`)).json()).comments,
-    ).toHaveLength(2)
-  })
-
-  it("resolves threads on republish via the resolves field", async () => {
-    const sid = (await (await upload("r.md", "# r", {})).json()).short_id
-    const cm = await (
-      await app.request(`/v1/artifacts/${sid}/comments`, json({ body_md: "fix this" }))
-    ).json()
-
-    const form = new FormData()
-    form.append("file", new Blob([new TextEncoder().encode("# r2")]), "r.md")
-    form.append("message", "address review")
-    form.append("resolves", cm.id)
-    await app.request(`/v1/artifacts/${sid}/versions`, { method: "POST", body: form })
-
-    expect(
-      (await (await app.request(`/v1/artifacts/${sid}/comments?state=open`)).json()).comments,
-    ).toHaveLength(0)
-  })
-
-  it("validates body and 404s unknown artifacts", async () => {
-    expect((await app.request(`/v1/artifacts/${shortId}/comments`, json({}))).status).toBe(400)
-    expect((await app.request("/v1/artifacts/zzzzzzzz/comments")).status).toBe(404)
-  })
-})
-
-describe("auth: token write-gating + per-artifact read-gating", () => {
-  const authApp = createApp({
-    meta,
-    blobs: new FsBlobStore(join(dir, "blobs")),
-    baseUrl: "http://dock.test",
-    token: "s3cret",
-  })
-  const authed = (extra: RequestInit = {}) => ({
-    ...extra,
-    headers: { authorization: "Bearer s3cret", ...(extra.headers ?? {}) },
-  })
-  const pub = (visibility?: string, headers?: HeadersInit) => {
-    const form = new FormData()
-    form.append("file", new Blob([new TextEncoder().encode("<h1>secret</h1>")]), "s.html")
-    if (visibility) form.append("visibility", visibility)
-    return authApp.request("/v1/artifacts", { method: "POST", body: form, headers })
-  }
-
-  it("rejects writes without the token, accepts with it", async () => {
-    expect((await pub("link")).status).toBe(403) // permission-gated: forbidden
-    const ok = await pub("link", { authorization: "Bearer s3cret" })
-    expect(ok.status).toBe(201)
-  })
-
-  it("serves public/link artifacts to anyone", async () => {
-    const { short_id } = await (await pub("link", { authorization: "Bearer s3cret" })).json()
-    expect((await authApp.request(`/a/${short_id}`)).status).toBe(200)
-    expect((await authApp.request(`/v1/artifacts/${short_id}/content`)).status).toBe(200)
-  })
-
-  it("hides gated artifacts without the token (404), reveals with it", async () => {
-    const { short_id } = await (await pub("org", { authorization: "Bearer s3cret" })).json()
-    expect((await authApp.request(`/a/${short_id}`)).status).toBe(404)
-    expect((await authApp.request(`/v1/artifacts/${short_id}`)).status).toBe(404)
-    expect((await authApp.request(`/v1/artifacts/${short_id}/content`)).status).toBe(404)
-    expect((await authApp.request(`/a/${short_id}`, authed())).status).toBe(200)
-    expect((await authApp.request(`/v1/artifacts/${short_id}`, authed())).status).toBe(200)
-  })
-
-  it("leaves a no-token instance fully open (the default app)", async () => {
-    const form = new FormData()
-    form.append("file", new Blob([new TextEncoder().encode("# x")]), "x.md")
-    expect((await app.request("/v1/artifacts", { method: "POST", body: form })).status).toBe(201)
-  })
-})
-
-describe("live stream (SSE)", () => {
-  async function readUntil(
-    reader: ReadableStreamDefaultReader<Uint8Array>,
-    needle: string,
-    timeoutMs = 2500,
-  ): Promise<string> {
-    const dec = new TextDecoder()
-    let buf = ""
-    const start = Date.now()
-    while (Date.now() - start < timeoutMs) {
-      const res = await Promise.race([
-        reader.read(),
-        new Promise<{ value?: Uint8Array; done: boolean }>((r) =>
-          setTimeout(() => r({ value: undefined, done: false }), 150),
-        ),
-      ])
-      if (res.value) buf += dec.decode(res.value, { stream: true })
-      if (buf.includes(needle)) return buf
-      if (res.done) break
-    }
-    throw new Error(`SSE timeout waiting for "${needle}"; got:\n${buf}`)
-  }
-
-  it("emits ready, then comment.created and version.published", async () => {
-    const { short_id } = await (await upload("live.md", "# live", {})).json()
-    const res = await app.request(`/v1/artifacts/${short_id}/events`)
-    expect(res.headers.get("content-type")).toContain("text/event-stream")
-    const reader = res.body!.getReader()
-    try {
-      await readUntil(reader, "event: ready")
-
-      await app.request(`/v1/artifacts/${short_id}/comments`, json({ body_md: "live note" }))
-      expect(await readUntil(reader, "event: comment.created")).toContain("live note")
-
-      const form = new FormData()
-      form.append("file", new Blob([new TextEncoder().encode("# live v2")]), "live.md")
-      form.append("message", "v2")
-      await app.request(`/v1/artifacts/${short_id}/versions`, { method: "POST", body: form })
-      expect(await readUntil(reader, "event: version.published")).toContain('"n":2')
-    } finally {
-      await reader.cancel()
-    }
-  })
-
-  it("reports presence on heartbeat", async () => {
-    const { short_id } = await (await upload("p.md", "# p", {})).json()
-    const res = await app.request(`/v1/artifacts/${short_id}/presence`, json({ name: "Jess" }))
-    expect((await res.json()).viewers).toEqual(["Jess"])
-  })
-})
-
-// A faithful role test needs real sessions. We seed Better Auth's `user` table
-// (so the share route can resolve email→id) and stand in a fake `auth` whose
-// session is chosen by an `x-test-user` header (the user's email). A static
-// token keeps the instance secured, so an unauthenticated caller is NOT owner.
-type TestUser = { id: string; email: string; name: string | null; image?: string | null }
-
-const makeAuthedApp = (
-  name: string,
-  users: TestUser[],
-  defaultRole?: AppDeps["defaultRole"],
-  multiWorkspace?: boolean,
-) => {
-  const path = join(dir, `${name}.db`)
-  const m = new SqliteMetaStore(path)
-  const raw = new Database(path)
-  raw.exec(
-    `CREATE TABLE IF NOT EXISTS user (id TEXT PRIMARY KEY, email TEXT, name TEXT, image TEXT)`,
-  )
-  const ins = raw.prepare(`INSERT OR IGNORE INTO user (id, email, name, image) VALUES (?,?,?,?)`)
-  for (const u of users) ins.run(u.id, u.email, u.name, u.image ?? null)
-  raw.close()
-  const auth = {
-    handler: async () => new Response(null, { status: 404 }),
-    api: {
-      getSession: async ({ headers }: { headers: Headers }) => {
-        const u = users.find((x) => x.email === headers.get("x-test-user"))
-        return u ? { user: u } : null
-      },
-    },
-  } as unknown as AppDeps["auth"]
-  const app = createApp({
-    meta: m,
-    blobs: new FsBlobStore(join(dir, "blobs")),
-    baseUrl: "http://dock.test",
-    token: "tok",
-    auth,
-    defaultRole,
-    multiWorkspace,
-    defaultOrgId: "default",
-  })
-  return { app, meta: m }
-}
-
-const as = (email: string) => ({ "x-test-user": email })
-const publishAs = (
-  app: ReturnType<typeof createApp>,
-  content: string,
-  fields: Record<string, string> = {},
-  headers: Record<string, string> = {},
-  shortId?: string,
-) => {
-  const form = new FormData()
-  form.append("file", new Blob([new TextEncoder().encode(content)]), "f.html")
-  for (const [k, v] of Object.entries(fields)) form.append(k, v)
-  const url = shortId ? `/v1/artifacts/${shortId}/versions` : "/v1/artifacts"
-  return app.request(url, { method: "POST", body: form, headers })
-}
-const jsonAs = (headers: Record<string, string>, body: unknown) => ({
-  method: "POST" as const,
-  headers: { "content-type": "application/json", ...headers },
-  body: JSON.stringify(body),
-})
-
-describe("permissions: workspace roles gate writes", () => {
-  const alice: TestUser = { id: "u_alice", email: "alice@dock.test", name: "Alice" }
-  const bob: TestUser = { id: "u_bob", email: "bob@dock.test", name: "Bob" }
-  const { app } = makeAuthedApp("perm-roles", [alice, bob], "commenter")
-  let shortId: string
-
-  it("makes the first member the owner, who can create artifacts", async () => {
-    const res = await publishAs(app, "<h1>a</h1>", {}, as(alice.email))
-    expect(res.status).toBe(201)
-    shortId = (await res.json()).short_id
-  })
-
-  it("provisions later members at the default role (commenter)", async () => {
-    const me = await (await app.request("/v1/me", { headers: as(bob.email) })).json()
-    expect(me.user.role).toBe("commenter")
-  })
-
-  it("blocks a commenter from creating or republishing", async () => {
-    expect((await publishAs(app, "<h1>b</h1>", {}, as(bob.email))).status).toBe(403)
-    expect((await publishAs(app, "<h1>a2</h1>", {}, as(bob.email), shortId)).status).toBe(403)
-  })
-
-  it("lets a commenter comment, but not an unauthenticated caller", async () => {
-    const ok = await app.request(
-      `/v1/artifacts/${shortId}/comments`,
-      jsonAs(as(bob.email), { body_md: "nice" }),
-    )
-    expect(ok.status).toBe(201)
-    const anon = await app.request(
-      `/v1/artifacts/${shortId}/comments`,
-      jsonAs({}, { body_md: "x" }),
-    )
-    expect(anon.status).toBe(403)
-  })
-})
-
-describe("permissions: a per-artifact share overrides the workspace role", () => {
-  const owner: TestUser = { id: "u_own", email: "own@dock.test", name: "Own" }
-  const carol: TestUser = { id: "u_carol", email: "carol@dock.test", name: "Carol" }
-  const { app } = makeAuthedApp("perm-share", [owner, carol], "viewer")
-  let shortId: string
-
-  it("a workspace viewer can read an org artifact but not republish it", async () => {
-    shortId = (
-      await (await publishAs(app, "<h1>secret</h1>", { visibility: "org" }, as(owner.email))).json()
-    ).short_id
-    expect(
-      (await app.request(`/v1/artifacts/${shortId}`, { headers: as(carol.email) })).status,
-    ).toBe(200)
-    expect((await publishAs(app, "<h1>edit</h1>", {}, as(carol.email), shortId)).status).toBe(403)
-  })
-
-  it("sharing the artifact as editor lets the viewer republish, and my_role reflects it", async () => {
-    const share = await app.request(`/v1/artifacts/${shortId}/members`, {
-      method: "PUT",
-      headers: { "content-type": "application/json", ...as(owner.email) },
-      body: JSON.stringify({ email: carol.email, role: "editor" }),
-    })
-    expect(share.status).toBe(201)
-    expect((await publishAs(app, "<h1>edit</h1>", {}, as(carol.email), shortId)).status).toBe(201)
-    const meta = await (
-      await app.request(`/v1/artifacts/${shortId}`, { headers: as(carol.email) })
-    ).json()
-    expect(meta.my_role).toBe("editor")
-  })
-
-  it("only an owner can manage shares (an editor cannot)", async () => {
-    const byEditor = await app.request(`/v1/artifacts/${shortId}/members`, {
-      method: "PUT",
-      headers: { "content-type": "application/json", ...as(carol.email) },
-      body: JSON.stringify({ email: owner.email, role: "viewer" }),
-    })
-    expect(byEditor.status).toBe(403)
-  })
-
-  it("the owner sees the share in the member list", async () => {
-    const list = await (
-      await app.request(`/v1/artifacts/${shortId}/members`, { headers: as(owner.email) })
-    ).json()
-    expect(list.default_role).toBe("viewer")
-    expect(list.members).toContainEqual(
-      expect.objectContaining({ email: carol.email, role: "editor" }),
-    )
-  })
-})
-
-const proposeAs = (
-  app: ReturnType<typeof createApp>,
-  shortId: string,
-  content: string,
-  headers: Record<string, string> = {},
-  fields: Record<string, string> = {},
-) => {
-  const form = new FormData()
-  form.append("file", new Blob([new TextEncoder().encode(content)]), "f.html")
-  for (const [k, v] of Object.entries(fields)) form.append(k, v)
-  return app.request(`/v1/artifacts/${shortId}/proposals`, { method: "POST", body: form, headers })
-}
-
-const bearer = (token: string) => ({ authorization: `Bearer ${token}` })
-
-describe("agents: @mention → pull inbox → propose → ack", () => {
-  const owner: TestUser = { id: "u_ag_own", email: "agown@dock.test", name: "Owner" }
-  const dev: TestUser = { id: "u_ag_dev", email: "agdev@dock.test", name: "Dev" }
-  const { app } = makeAuthedApp("agents", [owner, dev], "commenter")
-  let shortId: string
-  let agentId: string
-  let agentToken: string
-  let mentionId: string
-
-  it("an owner registers an agent and gets its token once; a commenter cannot", async () => {
-    // Provision roles in order: owner first (claims ownership), then dev (commenter).
-    await app.request("/v1/me", { headers: as(owner.email) })
-    await app.request("/v1/me", { headers: as(dev.email) })
-    expect((await app.request("/v1/agents", jsonAs(as(dev.email), { name: "Nope" }))).status).toBe(
-      403,
-    )
-    const res = await app.request("/v1/agents", jsonAs(as(owner.email), { name: "Claude" }))
-    expect(res.status).toBe(201)
-    const a = await res.json()
-    agentId = a.id
-    agentToken = a.token
-    expect(a.role).toBe("commenter")
-    expect(typeof agentToken).toBe("string")
-  })
-
-  it("stores only the token hash at rest, yet the raw token still authenticates", async () => {
-    const ag: TestUser = { id: "u_ag_hash", email: "aghash@dock.test", name: "Hasher" }
-    const { app: a2, meta: m2 } = makeAuthedApp("agent-hash", [ag], "commenter")
-    await a2.request("/v1/me", { headers: as(ag.email) }) // claims ownership
-    const reg = await (await a2.request("/v1/agents", jsonAs(as(ag.email), { name: "Bot" }))).json()
-    const raw = reg.token as string
-    // Single-workspace mode provisions under the default org id.
-    const [stored] = await m2.listAgents("default")
-    expect(stored.token).not.toBe(raw) // the raw secret is never at rest
-    expect(stored.token).toMatch(/^[0-9a-f]{64}$/) // it's a sha-256 hash
-    // The raw token still works: the agent can read its own inbox.
-    const inbox = await a2.request("/v1/agent/inbox", { headers: bearer(raw) })
-    expect(inbox.status).toBe(200)
-    // A near-miss token (its hash) does not authenticate.
-    const wrong = await a2.request("/v1/agent/inbox", { headers: bearer(stored.token) })
-    expect(wrong.status).toBe(401)
-    m2.close()
-  })
-
-  it("the agent shows up in the @mention directory", async () => {
-    const dir = await (await app.request("/v1/users?q=clau", { headers: as(owner.email) })).json()
-    expect(dir.users).toContainEqual(
-      expect.objectContaining({ id: agentId, name: "Claude", kind: "agent" }),
-    )
-  })
-
-  it("a comment @mentioning the agent lands in the agent's inbox, not a notification", async () => {
-    shortId = (await (await publishAs(app, "<h1>draft</h1>", {}, as(owner.email))).json()).short_id
-    const cm = await app.request(
-      `/v1/artifacts/${shortId}/comments`,
-      jsonAs(as(owner.email), {
-        body_md: "@Claude tighten the headline",
-        mentions: [{ id: agentId, name: "Claude" }],
-      }),
-    )
-    expect(cm.status).toBe(201)
-
-    // The agent pulls its inbox with its own token.
-    const inbox = await (
-      await app.request("/v1/agent/inbox", { headers: bearer(agentToken) })
-    ).json()
-    expect(inbox.agent.id).toBe(agentId)
-    expect(inbox.mentions).toHaveLength(1)
-    expect(inbox.mentions[0].artifact).toBe(shortId)
-    expect(inbox.mentions[0].body).toContain("tighten the headline")
-    mentionId = inbox.mentions[0].id
-  })
-
-  it("the agent proposes a candidate (commenter rank) authored as itself", async () => {
-    const res = await proposeAs(app, shortId, "<h1>A tighter headline</h1>", bearer(agentToken), {
-      message: "tightened per the mention",
-    })
-    expect(res.status).toBe(201)
-    const p = await res.json()
-    expect(p.author).toBe("Claude")
-    expect(p.state).toBe("open")
-    // It did NOT go live — a human still approves.
-    const art = await (
-      await app.request(`/v1/artifacts/${shortId}`, { headers: as(owner.email) })
-    ).json()
-    expect(art.current_version).toBe(1)
-    expect(art.open_proposals).toBe(1)
-  })
-
-  it("the agent cannot approve its own proposal (commenter, not editor)", async () => {
-    const open = await (
-      await app.request(`/v1/artifacts/${shortId}/proposals?state=open`, {
-        headers: as(owner.email),
-      })
-    ).json()
-    const pid = open.proposals[0].id
-    const res = await app.request(`/v1/artifacts/${shortId}/proposals/${pid}/approve`, {
-      method: "POST",
-      headers: bearer(agentToken),
-    })
-    expect(res.status).toBe(403)
-  })
-
-  it("acking the mention clears it from the inbox", async () => {
-    const ack = await app.request(`/v1/agent/mentions/${mentionId}/ack`, {
-      method: "POST",
-      headers: bearer(agentToken),
-    })
-    expect(ack.status).toBe(200)
-    const inbox = await (
-      await app.request("/v1/agent/inbox", { headers: bearer(agentToken) })
-    ).json()
-    expect(inbox.mentions).toHaveLength(0)
-  })
-
-  it("the inbox rejects a non-agent caller", async () => {
-    expect((await app.request("/v1/agent/inbox", { headers: as(owner.email) })).status).toBe(401)
-  })
-})
-
-describe("reviews: propose → approve goes live; commenter can't approve", () => {
-  const owner: TestUser = { id: "u_ro", email: "ro@dock.test", name: "Ro" }
-  const cassie: TestUser = { id: "u_cassie", email: "cassie@dock.test", name: "Cassie" }
-  const { app } = makeAuthedApp("reviews", [owner, cassie], "commenter")
-  let shortId: string
-  let proposalId: string
-
-  it("owner publishes v1; a commenter proposes a candidate that does NOT go live", async () => {
-    shortId = (await (await publishAs(app, "<h1>v1 live</h1>", {}, as(owner.email))).json())
-      .short_id
-    const res = await proposeAs(app, shortId, "<h1>candidate</h1>", as(cassie.email), {
-      message: "tighten the headline",
-    })
-    expect(res.status).toBe(201)
-    const p = await res.json()
-    proposalId = p.id
-    expect(p.state).toBe("open")
-    expect(p.base_version).toBe(1)
-
-    // The artifact is untouched: still v1, but the review queue shows 1.
-    const art = await (
-      await app.request(`/v1/artifacts/${shortId}`, { headers: as(owner.email) })
-    ).json()
-    expect(art.current_version).toBe(1)
-    expect(art.open_proposals).toBe(1)
-  })
-
-  it("renders the proposed experience at its preview URL, distinct from current", async () => {
-    const proposed = await app.request(`/raw/${shortId}/p/${proposalId}/index.html`, {
-      headers: as(owner.email),
-    })
-    expect(proposed.status).toBe(200)
-    expect(await proposed.text()).toContain("<h1>candidate</h1>")
-    // The live content is still v1.
-    const live = await app.request(`/v1/artifacts/${shortId}/content`, { headers: as(owner.email) })
-    expect(await live.text()).toContain("v1 live")
-  })
-
-  it("a commenter cannot approve their own proposal", async () => {
-    const res = await app.request(`/v1/artifacts/${shortId}/proposals/${proposalId}/approve`, {
-      method: "POST",
-      headers: as(cassie.email),
-    })
-    expect(res.status).toBe(403)
-  })
-
-  it("an editor/owner approves: the proposed content becomes the new current version", async () => {
-    const res = await app.request(`/v1/artifacts/${shortId}/proposals/${proposalId}/approve`, {
-      method: "POST",
-      headers: as(owner.email),
-    })
-    expect(res.status).toBe(200)
-    const body = await res.json()
-    expect(body.state).toBe("approved")
-    expect(body.published).toBe(2)
-
-    const art = await (
-      await app.request(`/v1/artifacts/${shortId}`, { headers: as(owner.email) })
-    ).json()
-    expect(art.current_version).toBe(2)
-    expect(art.open_proposals).toBe(0)
-    const live = await app.request(`/v1/artifacts/${shortId}/content`, { headers: as(owner.email) })
-    expect(await live.text()).toContain("candidate")
-  })
-
-  it("approving an already-decided proposal is a conflict", async () => {
-    const res = await app.request(`/v1/artifacts/${shortId}/proposals/${proposalId}/approve`, {
-      method: "POST",
-      headers: as(owner.email),
-    })
-    expect(res.status).toBe(409)
-  })
-})
-
-describe("reviews: request changes and withdraw keep content live-unchanged", () => {
-  const owner: TestUser = { id: "u_rc", email: "rc@dock.test", name: "Rc" }
-  const dana: TestUser = { id: "u_dana", email: "dana@dock.test", name: "Dana" }
-  const { app } = makeAuthedApp("reviews-rc", [owner, dana], "commenter")
-  let shortId: string
-
-  it("request-changes carries the reviewer's note back to the proposer, content unchanged", async () => {
-    shortId = (await (await publishAs(app, "<h1>base</h1>", {}, as(owner.email))).json()).short_id
-    const pid = (await (await proposeAs(app, shortId, "<h1>try</h1>", as(dana.email))).json()).id
-    const res = await app.request(`/v1/artifacts/${shortId}/proposals/${pid}/request-changes`, {
-      method: "POST",
-      headers: { "content-type": "application/json", ...as(owner.email) },
-      body: JSON.stringify({ note: "Tighten the intro paragraph first" }),
-    })
-    expect(res.status).toBe(200)
-    const decided = await res.json()
-    expect(decided.state).toBe("changes_requested")
-    expect(decided.decision_note).toBe("Tighten the intro paragraph first")
-    const art = await (
-      await app.request(`/v1/artifacts/${shortId}`, { headers: as(owner.email) })
-    ).json()
-    expect(art.current_version).toBe(1)
-    expect(art.open_proposals).toBe(0)
-    // The decided proposal still counts toward the Proposals entry (not withdrawn),
-    // so the proposer can return to read the feedback.
-    expect(art.proposals_total).toBe(1)
-  })
-
-  it("a proposer can withdraw their own open proposal", async () => {
-    const pid = (await (await proposeAs(app, shortId, "<h1>wip</h1>", as(dana.email))).json()).id
-    const res = await app.request(`/v1/artifacts/${shortId}/proposals/${pid}/withdraw`, {
-      method: "POST",
-      headers: as(dana.email),
-    })
-    expect(res.status).toBe(200)
-    expect((await res.json()).state).toBe("withdrawn")
-  })
-
-  it("an unauthenticated caller on a secured instance cannot propose", async () => {
-    const res = await proposeAs(app, shortId, "<h1>nope</h1>", {})
-    expect(res.status).toBe(403)
-  })
-})
-
-describe("anchored comments", () => {
-  it("flags comments anchored vs orphaned against the current version", async () => {
-    const sid = (await (await upload("a.md", "alpha beta gamma", { title: "A" })).json()).short_id
-    // anchor a comment to "beta"
-    const anchor = { type: "TextQuoteSelector", exact: "beta", prefix: "alpha ", suffix: " gamma" }
-    await app.request(`/v1/artifacts/${sid}/comments`, json({ body_md: "on beta", anchor }))
-
-    let list = await (await app.request(`/v1/artifacts/${sid}/comments`)).json()
-    expect(list.comments[0].anchored).toBe(true)
-
-    // republish without "beta" → comment becomes orphaned
-    const fd = new FormData()
-    fd.append("file", new Blob([new TextEncoder().encode("alpha gamma delta")]), "a.md")
-    fd.append("message", "v2")
-    await app.request(`/v1/artifacts/${sid}/versions`, { method: "POST", body: fd })
-
-    list = await (await app.request(`/v1/artifacts/${sid}/comments`)).json()
-    expect(list.comments[0].anchored).toBe(false)
-  })
-})
-
-describe("moderation: report → takedown (410) → reinstate + audit (C4a)", () => {
-  const owner: TestUser = { id: "u_mod_own", email: "modown@dock.test", name: "Owner" }
-  const dev: TestUser = { id: "u_mod_dev", email: "moddev@dock.test", name: "Dev" }
-  const { app } = makeAuthedApp("moderation", [owner, dev], "commenter")
-  let shortId: string
-
-  it("anyone can report a public artifact; reason is required", async () => {
-    await app.request("/v1/me", { headers: as(owner.email) })
-    await app.request("/v1/me", { headers: as(dev.email) })
-    shortId = (await (await publishAs(app, "<h1>spammy</h1>", {}, as(owner.email))).json()).short_id
-    // No reason → 400. Anonymous (no session) → still allowed.
-    expect((await app.request(`/v1/artifacts/${shortId}/report`, jsonAs({}, {}))).status).toBe(400)
-    const r = await app.request(
-      `/v1/artifacts/${shortId}/report`,
-      jsonAs({}, { reason: "phishing", detail: "fake login page" }),
-    )
-    expect(r.status).toBe(201)
-  })
-
-  it("the report shows in the owner's queue; a commenter can't see it", async () => {
-    expect((await app.request("/v1/reports", { headers: as(dev.email) })).status).toBe(403)
-    const q = await (await app.request("/v1/reports", { headers: as(owner.email) })).json()
-    expect(q.open).toBe(1)
-    expect(q.reports[0]).toMatchObject({ artifact_short_id: shortId, reason: "phishing" })
-  })
-
-  it("takedown 410s the content everywhere, keeps the record, and clears the report", async () => {
-    expect(
-      (await app.request(`/v1/artifacts/${shortId}/takedown`, jsonAs(as(dev.email), {}))).status,
-    ).toBe(403)
-    const td = await app.request(`/v1/artifacts/${shortId}/takedown`, jsonAs(as(owner.email), {}))
-    expect(td.status).toBe(200)
-
-    // Content is gone (410) on every serving surface; the record survives.
-    expect((await app.request(`/v1/artifacts/${shortId}/content`)).status).toBe(410)
-    expect((await app.request(`/raw/${shortId}/v/1/index.html`)).status).toBe(410)
-    expect((await app.request(`/a/${shortId}`)).status).toBe(410)
-    const art = await (
-      await app.request(`/v1/artifacts/${shortId}`, { headers: as(owner.email) })
-    ).json()
-    expect(art.removed).toBe(true)
-    // The open report was actioned by the takedown.
-    expect(
-      (await (await app.request("/v1/reports", { headers: as(owner.email) })).json()).open,
-    ).toBe(0)
-  })
-
-  it("reinstate restores serving", async () => {
-    const re = await app.request(`/v1/artifacts/${shortId}/reinstate`, jsonAs(as(owner.email), {}))
-    expect(re.status).toBe(200)
-    expect((await app.request(`/raw/${shortId}/v/1/index.html`)).status).toBe(200)
-  })
-
-  it("the audit log records every moderation action", async () => {
-    const { audit } = await (await app.request("/v1/audit", { headers: as(owner.email) })).json()
-    const actions = audit.map((a: { action: string }) => a.action)
-    expect(actions).toContain("report")
-    expect(actions).toContain("takedown")
-    expect(actions).toContain("reinstate")
-  })
-})
-
-describe("security: sandbox serving origin (A4)", () => {
-  const app = createApp({
-    meta: new SqliteMetaStore(join(dir, "sandbox.db")),
-    blobs: new FsBlobStore(join(dir, "blobs")),
-    baseUrl: "http://app.test",
-    sandboxOrigin: "http://sandbox.test",
-  })
-  let shortId: string
-
-  it("serves artifact bytes only from the sandbox host; the app host redirects there", async () => {
-    const fd = new FormData()
-    fd.append("file", new Blob([new TextEncoder().encode("<h1>user html</h1>")]), "x.html")
-    shortId = (
-      await (await app.request("http://app.test/v1/artifacts", { method: "POST", body: fd })).json()
-    ).short_id
-
-    // App (cookie) origin must never render user HTML — it 302s to the sandbox.
-    const onApp = await app.request(`http://app.test/raw/${shortId}/v/1/index.html`)
-    expect(onApp.status).toBe(302)
-    expect(onApp.headers.get("location")).toBe(`http://sandbox.test/raw/${shortId}/v/1/index.html`)
-
-    // The sandbox origin serves the bytes.
-    const onSandbox = await app.request(`http://sandbox.test/raw/${shortId}/v/1/index.html`)
-    expect(onSandbox.status).toBe(200)
-    expect(await onSandbox.text()).toContain("<h1>user html</h1>")
-  })
-
-  it("the sandbox host exposes ONLY raw bytes — never the API, auth, or the app", async () => {
-    expect((await app.request("http://sandbox.test/v1/artifacts")).status).toBe(404)
-    expect((await app.request("http://sandbox.test/api/auth/get-session")).status).toBe(404)
-    expect((await app.request(`http://sandbox.test/a/${shortId}`)).status).toBe(404)
-    // Health stays reachable on the sandbox host (for its own monitoring).
-    expect((await app.request("http://sandbox.test/healthz")).status).toBe(200)
-  })
-
-  it("the viewer shell points its iframe at the sandbox origin", async () => {
-    const shell = await (await app.request(`http://app.test/a/${shortId}`)).text()
-    expect(shell).toContain(`http://sandbox.test/raw/${shortId}/v/1/index.html`)
-  })
-
-  it("without a sandbox origin, raw is served in place (single-origin self-host)", async () => {
-    const solo = createApp({
-      meta: new SqliteMetaStore(join(dir, "solo.db")),
-      blobs: new FsBlobStore(join(dir, "blobs")),
-      baseUrl: "http://app.test",
-    })
-    const fd = new FormData()
-    fd.append("file", new Blob([new TextEncoder().encode("<h1>solo</h1>")]), "s.html")
-    const sid = (await (await solo.request("/v1/artifacts", { method: "POST", body: fd })).json())
-      .short_id
-    const raw = await solo.request(`/raw/${sid}/v/1/index.html`)
-    expect(raw.status).toBe(200)
-    expect(await raw.text()).toContain("<h1>solo</h1>")
-  })
-})
-
-describe("security: webhook SSRF guard", () => {
-  const owner: TestUser = { id: "u_wh", email: "wh@dock.test", name: "Wh" }
-  const { app } = makeAuthedApp("ssrf", [owner])
-  const create = (url: string) =>
-    app.request("/v1/webhooks", {
-      method: "POST",
-      headers: { ...as(owner.email), "content-type": "application/json" },
-      body: JSON.stringify({ url }),
-    })
-
-  it("rejects private, loopback, and metadata targets", async () => {
-    const blocked = [
-      "http://127.0.0.1/x",
-      "http://localhost/x",
-      "http://169.254.169.254/latest/meta-data",
-      "http://10.0.0.5/hook",
-      "http://192.168.1.10/hook",
-      "http://[::1]/x",
-      "ftp://example.com/x",
-    ]
-    for (const url of blocked) {
-      const r = await create(url)
-      expect(r.status, `should block ${url}`).toBe(400)
-    }
-  })
-
-  it("accepts a public https url", async () => {
-    expect((await create("https://hooks.example.com/abc")).status).toBe(201)
-  })
-})
-
-describe("security: rate limiting", () => {
-  const limited = createApp({
-    meta,
-    blobs: new FsBlobStore(join(dir, "blobs")),
-    baseUrl: "http://dock.test",
-    rateLimit: true,
-  })
-
-  it("429s once the auth window cap is exceeded", async () => {
-    let status = 0
-    for (let i = 0; i < 22; i++) {
-      status = (
-        await limited.request("/api/auth/get-session", {
-          headers: { "x-forwarded-for": "203.0.113.7" },
-        })
-      ).status
-    }
-    expect(status).toBe(429)
-  })
-})
-
-describe("@mentions + in-app notifications", () => {
-  const alice: TestUser = { id: "u_m_alice", email: "ma@dock.test", name: "Alice" }
-  const bob: TestUser = { id: "u_m_bob", email: "mb@dock.test", name: "Bob" }
-  const { app, meta: m } = makeAuthedApp("mentions", [alice, bob], "editor")
-  let shortId: string
-
-  it("lists provisioned workspace members in the mention directory, filtered by ?q=", async () => {
-    shortId = (await (await publishAs(app, "<h1>doc</h1>", {}, as(alice.email))).json()).short_id
-    await app.request("/v1/me", { headers: as(bob.email) }) // provisions bob as a member
-    const all = await (await app.request("/v1/users", { headers: as(alice.email) })).json()
-    const ids = all.users.map((u: { id: string }) => u.id)
-    expect(ids).toContain(alice.id)
-    expect(ids).toContain(bob.id)
-
-    const filtered = await (
-      await app.request("/v1/users?q=bob", { headers: as(alice.email) })
-    ).json()
-    expect(filtered.users).toHaveLength(1)
-    expect(filtered.users[0]).toMatchObject({ id: bob.id, name: "Bob", email: bob.email })
-  })
-
-  it("stores mentions on the comment and notifies the mentioned user, never the author", async () => {
-    const res = await app.request(
-      `/v1/artifacts/${shortId}/comments`,
-      jsonAs(as(alice.email), {
-        body_md: "hey please take a look",
-        mentions: [
-          { id: bob.id, name: "Bob" },
-          { id: alice.id, name: "Alice" }, // self-mention must NOT notify Alice
-        ],
-      }),
-    )
-    expect(res.status).toBe(201)
-    const cm = await res.json()
-    expect(cm.mentions).toEqual([
-      { id: bob.id, name: "Bob" },
-      { id: alice.id, name: "Alice" },
-    ])
-
-    const bobN = await (await app.request("/v1/notifications", { headers: as(bob.email) })).json()
-    expect(bobN.unread).toBe(1)
-    expect(bobN.notifications[0]).toMatchObject({
-      kind: "mention",
-      actor: "Alice",
-      artifact_short_id: shortId,
-      thread_id: cm.thread_id,
-      comment_id: cm.id,
-      read: 0,
-    })
-    expect(bobN.notifications[0].preview).toContain("take a look")
-
-    const aliceN = await (
-      await app.request("/v1/notifications", { headers: as(alice.email) })
-    ).json()
-    expect(aliceN.unread).toBe(0)
-  })
-
-  it("drops mentions of unknown user ids (no junk notifications)", async () => {
-    await app.request(
-      `/v1/artifacts/${shortId}/comments`,
-      jsonAs(as(alice.email), {
-        body_md: "ghost ping",
-        mentions: [{ id: "u_does_not_exist", name: "Ghost" }],
-      }),
-    )
-    // Bob still has only his single earlier notification; no row for the ghost id.
-    const ghost = await m.listNotifications("u_does_not_exist", 10)
-    expect(ghost).toHaveLength(0)
-  })
-
-  it("marks notifications read", async () => {
-    const before = await (await app.request("/v1/notifications", { headers: as(bob.email) })).json()
-    expect(before.unread).toBe(1)
-    const read = await app.request("/v1/notifications/read", jsonAs(as(bob.email), { all: true }))
-    expect((await read.json()).unread).toBe(0)
-    const after = await (await app.request("/v1/notifications", { headers: as(bob.email) })).json()
-    expect(after.unread).toBe(0)
-    expect(after.notifications[0].read).toBe(1)
-  })
-
-  it("requires auth for the directory and the notification feed", async () => {
-    expect((await app.request("/v1/users")).status).toBe(401)
-    expect((await app.request("/v1/notifications")).status).toBe(401)
-    expect((await app.request("/v1/notifications/events")).status).toBe(401)
-  })
-
-  it("enqueues a comment.mention webhook carrying the notified names", async () => {
-    await app.request("/v1/webhooks", {
-      method: "POST",
-      headers: { "content-type": "application/json", ...as(alice.email) },
-      body: JSON.stringify({
-        url: "https://hooks.example.com/mention",
-        events: ["comment.mention"],
-      }),
-    })
-    await app.request(
-      `/v1/artifacts/${shortId}/comments`,
-      jsonAs(as(alice.email), { body_md: "ping again", mentions: [{ id: bob.id, name: "Bob" }] }),
-    )
-    const due = await m.claimDueDeliveries(new Date(Date.now() + 1000).toISOString(), 50)
-    const mention = due.find((d) => d.event_type === "comment.mention")
-    expect(mention).toBeTruthy()
-    const payload = JSON.parse(mention?.payload ?? "{}")
-    expect(payload.data.mentioned).toContain("Bob")
-    expect(payload.data.author).toBe("Alice")
-  })
-
-  it("pushes a live notification event to the mentioned user's stream", async () => {
-    const res = await app.request("/v1/notifications/events", { headers: as(bob.email) })
-    expect(res.headers.get("content-type")).toContain("text/event-stream")
-    const reader = res.body?.getReader()
-    if (!reader) throw new Error("no stream body")
-    const dec = new TextDecoder()
-    const readUntil = async (needle: string, timeoutMs = 2500) => {
-      let buf = ""
-      const start = Date.now()
-      while (Date.now() - start < timeoutMs) {
-        const r = await Promise.race([
-          reader.read(),
-          new Promise<{ value?: Uint8Array; done: boolean }>((res) =>
-            setTimeout(() => res({ value: undefined, done: false }), 100),
-          ),
-        ])
-        if (r.value) buf += dec.decode(r.value, { stream: true })
-        if (buf.includes(needle)) return buf
-        if (r.done) break
-      }
-      throw new Error(`SSE timeout waiting for "${needle}"; got:\n${buf}`)
-    }
-    try {
-      await readUntil("event: ready")
-      await app.request(
-        `/v1/artifacts/${shortId}/comments`,
-        jsonAs(as(alice.email), { body_md: "live ping", mentions: [{ id: bob.id, name: "Bob" }] }),
-      )
-      const got = await readUntil("event: notification")
-      expect(got).toContain('"kind":"mention"')
-      expect(got).toContain("live ping")
-    } finally {
-      await reader.cancel()
-    }
-  })
-})
-
-describe("favorites + tags (browse)", () => {
-  const idOf = async (shortId: string) => {
-    const a = await meta.getByShortId(shortId)
-    if (!a) throw new Error(`no artifact ${shortId}`)
-    return a.id
-  }
-  const putTags = (shortId: string, tags: string[]) =>
-    app.request(`/v1/artifacts/${shortId}/tags`, {
-      method: "PUT",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({ tags }),
-    })
-
-  it("favorites are per-user at the store layer (set, idempotent, remove)", async () => {
-    const { short_id } = await (await upload("fav.md", "# Fav")).json()
-    const id = await idOf(short_id)
-    expect(await meta.listUserFavoriteIds("u1")).not.toContain(id)
-    await meta.setFavorite(id, "u1")
-    await meta.setFavorite(id, "u1") // idempotent — no duplicate row
-    expect(await meta.listUserFavoriteIds("u1")).toContain(id)
-    expect(await meta.listUserFavoriteIds("u2")).not.toContain(id) // personal
-    await meta.removeFavorite(id, "u1")
-    expect(await meta.listUserFavoriteIds("u1")).not.toContain(id)
-  })
-
-  it("normalizes tags and surfaces them on list + detail", async () => {
-    const { short_id } = await (await upload("t.md", "# Tagged")).json()
-    const res = await putTags(short_id, ["React", "react", "  Demo  ", ""])
-    expect(res.status).toBe(200)
-    expect((await res.json()).tags).toEqual(["demo", "react"]) // trimmed, lowercased, deduped, sorted
-
-    const detail = await (await app.request(`/v1/artifacts/${short_id}`)).json()
-    expect(detail.tags).toEqual(["demo", "react"])
-    expect(detail.favorite).toBe(false)
-
-    const list = await (await app.request("/v1/artifacts")).json()
-    const row = list.artifacts.find((x: { short_id: string }) => x.short_id === short_id)
-    expect(row.tags).toEqual(["demo", "react"])
-    expect(row).toHaveProperty("favorite")
-  })
-
-  it("replaces the full tag set (old tags drop)", async () => {
-    const { short_id } = await (await upload("r.md", "# R")).json()
-    await putTags(short_id, ["one", "two"])
-    expect((await (await putTags(short_id, ["three"])).json()).tags).toEqual(["three"])
-    const detail = await (await app.request(`/v1/artifacts/${short_id}`)).json()
-    expect(detail.tags).toEqual(["three"])
-  })
-
-  it("batches tags across artifacts without N+1", async () => {
-    const a1 = await (await upload("b1.md", "# 1")).json()
-    const a2 = await (await upload("b2.md", "# 2")).json()
-    await putTags(a1.short_id, ["solo"])
-    const id1 = await idOf(a1.short_id)
-    const id2 = await idOf(a2.short_id)
-    const map = await meta.tagsForArtifacts([id1, id2])
-    expect(map[id1]).toEqual(["solo"])
-    expect(map[id2]).toBeUndefined() // untagged ids simply have no entry
-  })
-})
-
-describe("server-side search + cursor pagination", () => {
-  const putTags = (shortId: string, tags: string[]) =>
-    app.request(`/v1/artifacts/${shortId}/tags`, {
-      method: "PUT",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({ tags }),
-    })
-
-  it("searches by title server-side, case-insensitive", async () => {
-    await upload("s1.md", "x", { title: "Quarterly ZZUNIQUE Report" })
-    await upload("s2.md", "x", { title: "Totally unrelated" })
-    const r = await (await app.request("/v1/artifacts?q=zzunique")).json()
-    const titles = r.artifacts.map((a: { title: string | null }) => a.title)
-    expect(titles).toContain("Quarterly ZZUNIQUE Report")
-    expect(titles.every((t: string | null) => /zzunique/i.test(t ?? ""))).toBe(true)
-  })
-
-  it("paginates newest-first with a keyset cursor (no overlap)", async () => {
-    for (const n of ["A", "B", "C"]) await upload(`pg${n}.md`, "x", { title: `PGSEED ${n}` })
-    const p1 = await (await app.request("/v1/artifacts?q=PGSEED&limit=2")).json()
-    expect(p1.artifacts).toHaveLength(2)
-    expect(typeof p1.next_cursor).toBe("string")
-    const p2 = await (
-      await app.request(
-        `/v1/artifacts?q=PGSEED&limit=2&cursor=${encodeURIComponent(p1.next_cursor)}`,
-      )
-    ).json()
-    expect(p2.artifacts).toHaveLength(1)
-    expect(p2.next_cursor).toBeNull()
-    const seen = new Set(p1.artifacts.map((a: { short_id: string }) => a.short_id))
-    expect(p2.artifacts.some((a: { short_id: string }) => seen.has(a.short_id))).toBe(false)
-  })
-
-  it("filters by ?tag= server-side", async () => {
-    const { short_id } = await (await upload("tg.md", "x", { title: "Tagged one" })).json()
-    await putTags(short_id, ["serverfilter"])
-    const r = await (await app.request("/v1/artifacts?tag=serverfilter")).json()
-    expect(r.artifacts.map((a: { short_id: string }) => a.short_id)).toEqual([short_id])
-  })
-
-  it("GET /v1/tags returns a browse summary (total, favorites, tag counts)", async () => {
-    const { short_id } = await (await upload("sum.md", "x", { title: "Summary doc" })).json()
-    await putTags(short_id, ["summaryfilter"])
-    const r = await (await app.request("/v1/tags")).json()
-    expect(typeof r.total).toBe("number")
-    expect(r.total).toBeGreaterThan(0)
-    expect(r.favorites).toBe(0) // anonymous in tests
-    expect(r.tags.find((t: { tag: string }) => t.tag === "summaryfilter")?.count).toBe(1)
-  })
-})
-
-describe("single-container web serving", () => {
-  it("serves the API landing at / by default", async () => {
-    const r = await app.request("/")
-    expect(r.status).toBe(200)
-    expect(await r.text()).toContain("open home for AI-generated artifacts")
-  })
-
-  it("drops the / placeholder when serveWeb is set, so the bundled SPA owns the shell", async () => {
-    const webApp = createApp({
-      meta,
-      blobs: new FsBlobStore(join(dir, "blobs")),
-      baseUrl: "http://dock.test",
-      serveWeb: true,
-    })
-    // No placeholder here; the Node entry's static + index.html fallback (added
-    // around createApp when a build is present) is what answers `/` in prod.
-    expect((await webApp.request("/")).status).toBe(404)
-  })
-})
-
-describe("collections", () => {
-  const put = (path: string) =>
-    app.request(path, { method: "PUT", headers: { "content-type": "application/json" } })
-
-  it("CRUD + items + ?collection filter + detail membership", async () => {
-    const col = await (await postJson("/v1/collections", { title: "Launch" })).json()
-    expect(col.title).toBe("Launch")
-    const { short_id } = await (await upload("c1.md", "x", { title: "In collection" })).json()
-    expect((await put(`/v1/collections/${col.id}/items/${short_id}`)).status).toBe(200)
-
-    const list = await (await app.request("/v1/collections")).json()
-    expect(list.collections.find((c: { id: string }) => c.id === col.id)?.count).toBe(1)
-
-    const filtered = await (await app.request(`/v1/artifacts?collection=${col.id}`)).json()
-    expect(filtered.artifacts.map((a: { short_id: string }) => a.short_id)).toEqual([short_id])
-
-    const detail = await (await app.request(`/v1/artifacts/${short_id}`)).json()
-    expect(detail.collections).toEqual([col.id])
-
-    await app.request(`/v1/collections/${col.id}/items/${short_id}`, { method: "DELETE" })
-    expect(
-      (await (await app.request(`/v1/artifacts?collection=${col.id}`)).json()).artifacts,
-    ).toHaveLength(0)
-  })
-
-  it("sharing a collection grants its role on every artifact in it", async () => {
-    const col = await (await postJson("/v1/collections", { title: "Shared" })).json()
-    const { short_id } = await (await upload("cs.md", "x", { title: "Shared art" })).json()
-    await put(`/v1/collections/${col.id}/items/${short_id}`)
-    const art = await meta.getByShortId(short_id)
-    if (!art) throw new Error("missing artifact")
-    // A stranger has no role over the artifact…
-    expect(await meta.collectionRolesForArtifact(art.id, "stranger")).toEqual([])
-    // …until the collection is shared with them — then it propagates to items.
-    await meta.setCollectionMember({
-      id: "cm_test",
-      collection_id: col.id,
-      user_id: "stranger",
-      role: "editor",
-    })
-    expect(await meta.collectionRolesForArtifact(art.id, "stranger")).toEqual(["editor"])
-    // Removing the item drops the propagation.
-    await meta.removeCollectionItem(col.id, art.id)
-    expect(await meta.collectionRolesForArtifact(art.id, "stranger")).toEqual([])
-  })
-
-  it("renames and deletes a collection (artifacts survive)", async () => {
-    const col = await (await postJson("/v1/collections", { title: "Temp" })).json()
-    const { short_id } = await (await upload("cd.md", "x", { title: "Survivor" })).json()
-    await put(`/v1/collections/${col.id}/items/${short_id}`)
-    await app.request(`/v1/collections/${col.id}`, {
-      method: "PATCH",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({ title: "Renamed" }),
-    })
-    expect((await meta.getCollection(col.id))?.title).toBe("Renamed")
-    expect((await app.request(`/v1/collections/${col.id}`, { method: "DELETE" })).status).toBe(204)
-    expect(await meta.getCollection(col.id)).toBeNull()
-    // the artifact itself is untouched
-    expect(await meta.getByShortId(short_id)).not.toBeNull()
-  })
-})
-
-describe("analytics: identity + retention", () => {
-  const ann: TestUser = { id: "u_view_ann", email: "ann@dock.test", name: "Ann" }
-  const bob: TestUser = {
-    id: "u_view_bob",
-    email: "bob@dock.test",
-    name: "Bob",
-    image: "https://cdn.dock.test/bob.png",
-  }
-  const { app, meta: m } = makeAuthedApp("analytics-id", [ann, bob], "commenter")
-
-  it("excludes the owner's own opens; counts a viewer (by name + avatar) and anon", async () => {
-    const sid = (await (await publishAs(app, "<h1>v</h1>", {}, as(ann.email))).json()).short_id
-    // Ann is the workspace owner (first member). Her own opens don't count.
-    for (let i = 0; i < 2; i++)
-      await app.request(`/v1/artifacts/${sid}/view`, jsonAs(as(ann.email), { version: 1 }))
-    // Bob is a commenter (audience); his two opens collapse to one counted view.
-    for (let i = 0; i < 2; i++)
-      await app.request(`/v1/artifacts/${sid}/view`, jsonAs(as(bob.email), { version: 1 }))
-    // An anonymous open also counts.
-    await app.request(`/v1/artifacts/${sid}/view`, {
-      method: "POST",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({ version: 1 }),
-    })
-    const a = await (
-      await app.request(`/v1/artifacts/${sid}/analytics`, { headers: as(ann.email) })
-    ).json()
-    expect(a.total).toBe(2) // Bob (de-duped) + anon; Ann excluded
-    expect(a.unique).toBe(2)
-    expect(a.anonViewers).toBe(1)
-    const named = a.recent.filter((r: { kind: string }) => r.kind === "user")
-    expect(named).toHaveLength(1)
-    expect(named[0]).toMatchObject({ viewer: "Bob", avatar: bob.image }) // name + profile pic
-    expect(a.recent.some((r: { viewer: string }) => r.viewer === "Ann")).toBe(false)
-  })
-
-  it("prunes view rows older than the cutoff, keeps newer ones", async () => {
-    const sid = (await (await publishAs(app, "<h1>p</h1>", {}, as(ann.email))).json()).short_id
-    const art = await m.getByShortId(sid)
-    if (!art) throw new Error("no artifact")
-    await m.recordView({
-      id: "v_old",
-      artifact_id: art.id,
-      version: 1,
-      viewer: "anon_p",
-      viewer_kind: "anon",
-    })
-    expect((await m.viewStats(art.id)).total).toBe(1)
-    expect(await m.pruneViews(new Date(Date.now() - 86400_000).toISOString())).toBe(0) // none that old
-    expect((await m.viewStats(art.id)).total).toBe(1) // kept
-    expect(await m.pruneViews(new Date(Date.now() + 86400_000).toISOString())).toBeGreaterThan(0)
-    expect((await m.viewStats(art.id)).total).toBe(0) // pruned
-  })
-
-  it("pruneViewsByViewers removes only the listed user-kind rows", async () => {
-    const sid = (await (await publishAs(app, "<h1>pv</h1>", {}, as(ann.email))).json()).short_id
-    const art = await m.getByShortId(sid)
-    if (!art) throw new Error("no artifact")
-    const v = (id: string, viewer: string, kind: "user" | "anon") =>
-      m.recordView({ id, artifact_id: art.id, version: 1, viewer, viewer_kind: kind })
-    await v("v_keep", "u_keep", "user")
-    await v("v_drop", "u_drop", "user")
-    await v("v_anon", "u_drop", "anon") // same string but anon-kind → must be kept
-    expect(await m.pruneViewsByViewers(["u_drop"])).toBe(1) // only the user-kind row
-    expect((await m.viewStats(art.id)).total).toBe(2) // u_keep + the anon row remain
-  })
-
-  it("the anonymous-viewer cookie is cross-site-safe when the SPA is split", async () => {
-    const xs = createApp({
-      meta,
-      blobs: new FsBlobStore(join(dir, "blobs")),
-      baseUrl: "https://api.dock.test",
-      crossSite: true,
-    })
-    const fd = new FormData()
-    fd.append("file", new Blob([new TextEncoder().encode("# x")]), "x.md")
-    const { short_id } = await (
-      await xs.request("/v1/artifacts", { method: "POST", body: fd })
-    ).json()
-    const r = await xs.request(`/v1/artifacts/${short_id}/view`, {
-      method: "POST",
-      headers: { "content-type": "application/json" },
-      body: "{}",
-    })
-    const sc = (r.headers.get("set-cookie") ?? "").toLowerCase()
-    expect(sc).toContain("dock_vid=")
-    expect(sc).toContain("samesite=none") // rides the cross-site request
-    expect(sc).toContain("secure")
-  })
-})
-
-describe("workspace: name + members (Admin / Creator / Viewer)", () => {
-  const admin: TestUser = { id: "u_ws_admin", email: "wsadmin@dock.test", name: "Ada" }
-  const creator: TestUser = { id: "u_ws_creator", email: "wscreator@dock.test", name: "Cara" }
-  const viewer: TestUser = { id: "u_ws_viewer", email: "wsviewer@dock.test", name: "Vic" }
-  const { app } = makeAuthedApp("workspace", [admin, creator, viewer], "commenter")
-
-  const ws = async (headers: Record<string, string>) =>
-    (await app.request("/v1/workspace", { headers })).json()
-  const patchWs = (headers: Record<string, string>, body: unknown) =>
-    app.request("/v1/workspace", { ...jsonAs(headers, body), method: "PATCH" })
-  const putMember = (headers: Record<string, string>, body: unknown) =>
-    app.request("/v1/workspace/members", { ...jsonAs(headers, body), method: "PUT" })
-  const patchMember = (headers: Record<string, string>, userId: string, body: unknown) =>
-    app.request(`/v1/workspace/members/${userId}`, { ...jsonAs(headers, body), method: "PATCH" })
-
-  it("defaults the name and makes the first user the Admin (owner)", async () => {
-    // The first /v1/me claims ownership; the others provision as Viewers.
-    const me = await (await app.request("/v1/me", { headers: as(admin.email) })).json()
-    expect(me.user.role).toBe("owner")
-    await app.request("/v1/me", { headers: as(creator.email) })
-    await app.request("/v1/me", { headers: as(viewer.email) })
-    const w = await ws(as(admin.email))
-    expect(w.name).toBe("My Workspace")
-    expect(w.role).toBe("owner")
-    expect(w.members).toHaveLength(3)
-  })
-
-  it("an Admin renames the workspace; a Viewer cannot; the name shows in browse", async () => {
-    const ok = await patchWs(as(admin.email), { name: "Acme HQ" })
-    expect(ok.status).toBe(200)
-    expect((await ok.json()).name).toBe("Acme HQ")
-    expect((await ws(as(viewer.email))).name).toBe("Acme HQ")
-    expect((await patchWs(as(viewer.email), { name: "Hijacked" })).status).toBe(403)
-    const summary = await (await app.request("/v1/tags", { headers: as(admin.email) })).json()
-    expect(summary.workspace).toBe("Acme HQ")
-  })
-
-  it("an Admin promotes a member to Creator (editor); a Viewer cannot add anyone", async () => {
-    expect(
-      (await putMember(as(admin.email), { email: creator.email, role: "editor" })).status,
-    ).toBe(201)
-    const me = await (await app.request("/v1/me", { headers: as(creator.email) })).json()
-    expect(me.user.role).toBe("editor")
-    expect((await putMember(as(viewer.email), { email: viewer.email, role: "owner" })).status).toBe(
-      403,
-    )
-  })
-
-  it("rejects an unknown email and an invalid role", async () => {
-    expect(
-      (await putMember(as(admin.email), { email: "ghost@dock.test", role: "editor" })).status,
-    ).toBe(404)
-    expect(
-      (await putMember(as(admin.email), { email: creator.email, role: "wizard" })).status,
-    ).toBe(400)
-  })
-
-  it("changes a member's role and removes a member", async () => {
-    expect((await patchMember(as(admin.email), creator.id, { role: "owner" })).status).toBe(200)
-    const del = await app.request(`/v1/workspace/members/${viewer.id}`, {
-      method: "DELETE",
-      headers: as(admin.email),
-    })
-    expect(del.status).toBe(204)
-    const w = await ws(as(admin.email))
-    expect(w.members.find((m: { user_id: string }) => m.user_id === viewer.id)).toBeUndefined()
-  })
-
-  it("won't strip or remove the last Admin", async () => {
-    // creator is currently an Admin too; demote it so `admin` is the only one.
-    expect((await patchMember(as(admin.email), creator.id, { role: "editor" })).status).toBe(200)
-    expect((await patchMember(as(admin.email), admin.id, { role: "editor" })).status).toBe(409)
-    const remove = await app.request(`/v1/workspace/members/${admin.id}`, {
-      method: "DELETE",
-      headers: as(admin.email),
-    })
-    expect(remove.status).toBe(409)
-  })
-})
-
-describe("workspace: edge conditions", () => {
-  const admin: TestUser = { id: "u_we_admin", email: "weadmin@dock.test", name: "Ed" }
-  const other: TestUser = { id: "u_we_other", email: "weother@dock.test", name: "Otto" }
-  const { app } = makeAuthedApp("ws-edge", [admin, other], "commenter")
-
-  const putMember = (headers: Record<string, string>, body: unknown) =>
-    app.request("/v1/workspace/members", { ...jsonAs(headers, body), method: "PUT" })
-  const patchWs = (headers: Record<string, string>, body: unknown) =>
-    app.request("/v1/workspace", { ...jsonAs(headers, body), method: "PATCH" })
-
-  it("provisions the first user as Admin, then clamps long names and rejects blank ones", async () => {
-    await app.request("/v1/me", { headers: as(admin.email) }) // first member → owner
-    expect((await patchWs(as(admin.email), { name: "   " })).status).toBe(400)
-    expect((await patchWs(as(admin.email), {})).status).toBe(400)
-    const ok = await patchWs(as(admin.email), { name: "x".repeat(200) })
-    expect(ok.status).toBe(200)
-    expect((await ok.json()).name).toHaveLength(80)
-  })
-
-  it("rejects any role outside Admin / Creator / Viewer", async () => {
-    for (const role of ["viewer", "admin", "boss", 3, null]) {
-      expect((await putMember(as(admin.email), { email: other.email, role })).status).toBe(400)
-    }
-  })
-
-  it("PUT is idempotent: re-adding a member re-roles them without duplicating", async () => {
-    expect(
-      (await putMember(as(admin.email), { email: other.email, role: "commenter" })).status,
-    ).toBe(201)
-    expect((await putMember(as(admin.email), { email: other.email, role: "editor" })).status).toBe(
-      201,
-    )
-    const w = await (await app.request("/v1/workspace", { headers: as(admin.email) })).json()
-    const rows = w.members.filter((m: { user_id: string }) => m.user_id === other.id)
-    expect(rows).toHaveLength(1)
-    expect(rows[0].role).toBe("editor")
-  })
-
-  it("the add/update (PUT) path also refuses to strip the last Admin", async () => {
-    // admin is the sole owner — demoting itself via PUT must be blocked too.
-    expect(
-      (await putMember(as(admin.email), { email: admin.email, role: "commenter" })).status,
-    ).toBe(409)
-  })
-
-  it("404s a PATCH on a non-member; a Creator can't manage", async () => {
-    // `other` is a Creator (editor) — no manage rights.
-    expect((await patchWs(as(other.email), { name: "Nope" })).status).toBe(403)
-    expect((await putMember(as(other.email), { email: other.email, role: "owner" })).status).toBe(
-      403,
-    )
-    const ghost = await app.request("/v1/workspace/members/u_ghost", {
-      ...jsonAs(as(admin.email), { role: "editor" }),
-      method: "PATCH",
-    })
-    expect(ghost.status).toBe(404)
-  })
-
-  it("allows self-demotion once a second Admin exists", async () => {
-    expect((await putMember(as(admin.email), { email: other.email, role: "owner" })).status).toBe(
-      201,
-    )
-    const demote = await app.request(`/v1/workspace/members/${admin.id}`, {
-      ...jsonAs(as(admin.email), { role: "editor" }),
-      method: "PATCH",
-    })
-    expect(demote.status).toBe(200)
-    // having dropped Admin, the former Admin can no longer manage.
-    expect((await patchWs(as(admin.email), { name: "X" })).status).toBe(403)
-  })
-})
-
-describe("workspace: open + token modes", () => {
-  it("an unsecured instance trusts anon as Admin and defaults the name", async () => {
-    const openApp = createApp({
-      meta: new SqliteMetaStore(join(dir, "ws-open.db")),
-      blobs: new FsBlobStore(join(dir, "blobs")),
-      baseUrl: "http://dock.test",
-    })
-    const w = await (await openApp.request("/v1/workspace")).json()
-    expect(w.name).toBe("My Workspace")
-    expect(w.role).toBe("owner")
-    expect(w.members).toEqual([])
-    const renamed = await openApp.request("/v1/workspace", {
-      method: "PATCH",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({ name: "Solo" }),
-    })
-    expect(renamed.status).toBe(200)
-    expect((await renamed.json()).name).toBe("Solo")
-  })
-
-  it("a static token acts as Admin", async () => {
-    const admin: TestUser = { id: "u_tok_admin", email: "tokadmin@dock.test", name: "Toki" }
-    const { app: tokApp } = makeAuthedApp("ws-token", [admin], "commenter")
-    const r = await tokApp.request("/v1/workspace", {
-      method: "PATCH",
-      headers: { "content-type": "application/json", ...bearer("tok") },
-      body: JSON.stringify({ name: "Tokenspace" }),
-    })
-    expect(r.status).toBe(200)
-    expect((await r.json()).name).toBe("Tokenspace")
-  })
-})
-
-describe("multi-workspace: isolation, switch, create, provision", () => {
-  const ada: TestUser = { id: "u_mw_ada", email: "mwada@dock.test", name: "Ada" }
-  const bo: TestUser = { id: "u_mw_bo", email: "mwbo@dock.test", name: "Bo" }
-  const { app } = makeAuthedApp("multiws", [ada, bo], "commenter", true)
-
-  // Pull the dock_ws cookie out of a Set-Cookie header to thread it on later calls.
-  const wsCookie = (res: Response): string => {
-    const m = (res.headers.get("set-cookie") ?? "").match(/dock_ws=([^;]+)/)
-    return m ? `dock_ws=${m[1]}` : ""
-  }
-  const withCookie = (email: string, cookie: string) => ({ ...as(email), cookie })
-
-  it("provisions a personal workspace for each new user on first load", async () => {
-    const a = await (await app.request("/v1/workspaces", { headers: as(ada.email) })).json()
-    expect(a.multi).toBe(true)
-    expect(a.workspaces).toHaveLength(1)
-    expect(a.workspaces[0].role).toBe("owner")
-    const b = await (await app.request("/v1/workspaces", { headers: as(bo.email) })).json()
-    // Bo gets his OWN workspace, distinct from Ada's.
-    expect(b.workspaces[0].id).not.toBe(a.workspaces[0].id)
-  })
-
-  it("isolates artifacts to the active workspace", async () => {
-    expect((await publishAs(app, "<h1>ada</h1>", { title: "Ada doc" }, as(ada.email))).status).toBe(
-      201,
-    )
-    const adaList = await (await app.request("/v1/artifacts", { headers: as(ada.email) })).json()
-    expect(adaList.artifacts.map((x: { title: string }) => x.title)).toContain("Ada doc")
-    const boList = await (await app.request("/v1/artifacts", { headers: as(bo.email) })).json()
-    expect(boList.artifacts).toHaveLength(0)
-  })
-
-  it("creates a workspace, switches into it, and it starts empty", async () => {
-    const create = await app.request("/v1/workspaces", jsonAs(as(ada.email), { name: "Acme" }))
-    expect(create.status).toBe(201)
-    const { id: acmeId, role } = await create.json()
-    expect(role).toBe("owner")
-    const cookie = wsCookie(create)
-    expect(cookie).toContain(acmeId)
-    // In Acme (via the cookie), Ada sees none of her personal-workspace artifacts.
-    const inAcme = await (
-      await app.request("/v1/artifacts", { headers: withCookie(ada.email, cookie) })
-    ).json()
-    expect(inAcme.artifacts).toHaveLength(0)
-    const spaces = await (
-      await app.request("/v1/workspaces", { headers: withCookie(ada.email, cookie) })
-    ).json()
-    expect(spaces.active).toBe(acmeId)
-    expect(spaces.workspaces).toHaveLength(2)
-  })
-
-  it("switches back to a workspace you belong to; rejects one you don't", async () => {
-    const a = await (await app.request("/v1/workspaces", { headers: as(ada.email) })).json()
-    const personal = a.workspaces.find((w: { name: string }) => w.name !== "Acme")
-    const sw = await app.request("/v1/workspace/switch", jsonAs(as(ada.email), { id: personal.id }))
-    expect(sw.status).toBe(200)
-    const back = await (
-      await app.request("/v1/artifacts", { headers: withCookie(ada.email, wsCookie(sw)) })
-    ).json()
-    expect(back.artifacts.map((x: { title: string }) => x.title)).toContain("Ada doc")
-    // Bo is not a member of Ada's personal workspace.
-    expect(
-      (await app.request("/v1/workspace/switch", jsonAs(as(bo.email), { id: personal.id }))).status,
-    ).toBe(403)
-  })
-})
-
-describe("multi-workspace: create/switch are disabled in single mode", () => {
-  const ada: TestUser = { id: "u_sw_ada", email: "swada@dock.test", name: "Ada" }
-  const { app } = makeAuthedApp("singlews", [ada], "commenter") // multi off
-
-  it("403s create and switch, and reports multi:false", async () => {
-    expect((await app.request("/v1/workspaces", jsonAs(as(ada.email), { name: "X" }))).status).toBe(
-      403,
-    )
-    expect(
-      (await app.request("/v1/workspace/switch", jsonAs(as(ada.email), { id: "whatever" }))).status,
-    ).toBe(403)
-    const list = await (await app.request("/v1/workspaces", { headers: as(ada.email) })).json()
-    expect(list.multi).toBe(false)
-    expect(list.workspaces).toHaveLength(1)
-  })
-})
-
-// ---------------------------------------------------------------------------
-// C4b — launch-gate prevention: storage quotas + per-actor rate limits.
-// ---------------------------------------------------------------------------
-
-// A fresh app with custom deps. Without `users` it's an open instance (anonymous
-// = owner) — handy for quota/anonymous-flood tests; with `users` it's secured
-// and the session is picked by an `x-test-user` header, for per-actor tests.
-const quotaApp = (name: string, extra: Partial<AppDeps>, users?: TestUser[]) => {
-  const path = join(dir, `${name}.db`)
-  const m = new SqliteMetaStore(path)
-  let auth: AppDeps["auth"]
-  if (users) {
-    const raw = new Database(path)
-    raw.exec(`CREATE TABLE IF NOT EXISTS user (id TEXT PRIMARY KEY, email TEXT, name TEXT)`)
-    const ins = raw.prepare(`INSERT OR IGNORE INTO user (id, email, name) VALUES (?,?,?)`)
-    for (const u of users) ins.run(u.id, u.email, u.name)
-    raw.close()
-    auth = {
-      handler: async () => new Response(null, { status: 404 }),
-      api: {
-        getSession: async ({ headers }: { headers: Headers }) => {
-          const u = users.find((x) => x.email === headers.get("x-test-user"))
-          return u ? { user: u } : null
-        },
-      },
-    } as unknown as AppDeps["auth"]
-  }
-  const app = createApp({
-    meta: m,
-    blobs: new FsBlobStore(join(dir, `blobs-${name}`)),
-    baseUrl: "http://dock.test",
-    ...(users ? { token: "tok", auth } : {}),
-    ...extra,
-  })
-  return { app, meta: m }
-}
-
-// Publish a string payload, optionally as a new version (shortId) and/or as a
-// given session (headers). Byte length = the string's UTF-8 length.
-const pub = (
-  app: ReturnType<typeof createApp>,
-  content: string,
-  fields: Record<string, string> = {},
-  shortId?: string,
-  headers: Record<string, string> = {},
-) => {
-  const form = new FormData()
-  form.append("file", new Blob([new TextEncoder().encode(content)]), "f.html")
-  for (const [k, v] of Object.entries(fields)) form.append(k, v)
-  const url = shortId ? `/v1/artifacts/${shortId}/versions` : "/v1/artifacts"
-  return app.request(url, { method: "POST", body: form, headers })
-}
-
-describe("quotas: per-workspace storage caps (C4b)", () => {
-  it("persists each version's byte size and meters the workspace total", async () => {
-    const { app, meta: m } = quotaApp("quota-meter", {})
-    const a = await (await pub(app, "hello")).json() // 5 bytes
-    await pub(app, "worldwide", {}, a.short_id) // +9 bytes on v2
-    expect(await m.storageBytes("default")).toBe(14)
-    // Republishing identical bytes is content-addressed to the same blob, so the
-    // meter dedups it — still 14, not 19.
-    await pub(app, "hello", {}, a.short_id)
-    expect(await m.storageBytes("default")).toBe(14)
-    m.close()
-  })
-
-  it("rejects an upload that would exceed maxBytes with 413, keeps the under-limit one", async () => {
-    const { app, meta: m } = quotaApp("quota-bytes", { maxBytes: 20 })
-    expect((await pub(app, "0123456789")).status).toBe(201) // 10 ≤ 20
-    const over = await pub(app, "0123456789AB") // total 22 > 20
-    expect(over.status).toBe(413)
-    expect((await over.json()).error).toMatch(/storage quota/)
-    expect(await m.storageBytes("default")).toBe(10) // the rejected upload stored nothing
-    m.close()
-  })
-
-  it("caps the number of artifacts (409), but still allows new versions of existing ones", async () => {
-    const { app } = quotaApp("quota-count", { maxArtifacts: 2 })
-    const a1 = await (await pub(app, "one")).json()
-    expect((await pub(app, "two")).status).toBe(201)
-    const third = await pub(app, "three")
-    expect(third.status).toBe(409)
-    expect((await third.json()).error).toMatch(/artifact quota/)
-    // Republishing an existing artifact doesn't create a new one — still allowed.
-    expect((await pub(app, "one-v2", {}, a1.short_id)).status).toBe(201)
-  })
-})
-
-describe("rate limits: per-actor write throttles (C4b)", () => {
-  it("throttles a flood of publishes from one caller with 429 + Retry-After", async () => {
-    const { app } = quotaApp("rl-publish", { rateLimit: true, publishRate: 2 })
-    expect((await pub(app, "a")).status).toBe(201)
-    expect((await pub(app, "b")).status).toBe(201)
-    const blocked = await pub(app, "c")
-    expect(blocked.status).toBe(429)
-    expect(blocked.headers.get("Retry-After")).toBeTruthy()
-  })
-
-  it("throttles comments independently of publishes", async () => {
-    const { app } = quotaApp("rl-comment", { rateLimit: true, commentRate: 2 })
-    const a = await (await pub(app, "doc")).json()
-    const comment = () =>
-      app.request(`/v1/artifacts/${a.short_id}/comments`, {
-        method: "POST",
-        headers: { "content-type": "application/json" },
-        body: JSON.stringify({ body_md: "hi", author: "x" }),
-      })
-    expect((await comment()).status).toBe(201)
-    expect((await comment()).status).toBe(201)
-    expect((await comment()).status).toBe(429)
-  })
-
-  it("keys the limit by identity — one user hitting the cap doesn't block another", async () => {
-    const amy: TestUser = { id: "u_amy", email: "amy@dock.test", name: "Amy" }
-    const ben: TestUser = { id: "u_ben", email: "ben@dock.test", name: "Ben" }
-    const { app } = quotaApp("rl-actor", { rateLimit: true, publishRate: 1 }, [amy, ben])
-    // Amy provisions as owner (first member); Ben joins as the default editor.
-    await app.request("/v1/me", { headers: as(amy.email) })
-    await app.request("/v1/me", { headers: as(ben.email) })
-    expect((await pub(app, "a1", {}, undefined, as(amy.email))).status).toBe(201)
-    expect((await pub(app, "a2", {}, undefined, as(amy.email))).status).toBe(429) // Amy capped
-    expect((await pub(app, "b1", {}, undefined, as(ben.email))).status).toBe(201) // Ben unaffected
-  })
-})
-
-describe("multi-tenant hardening: per-org quotas + cross-org isolation", () => {
-  const amy: TestUser = { id: "u_mt_amy", email: "mtamy@dock.test", name: "Amy" }
-  const ben: TestUser = { id: "u_mt_ben", email: "mtben@dock.test", name: "Ben" }
-
-  it("storage quota is per-workspace — one org filling up does NOT block another", async () => {
-    const { app } = quotaApp("mt-quota", { multiWorkspace: true, maxBytes: 12 }, [amy, ben])
-    await app.request("/v1/me", { headers: as(amy.email) }) // provision Amy's workspace
-    await app.request("/v1/me", { headers: as(ben.email) }) // provision Ben's workspace
-    // Amy fills her own 12-byte cap.
-    expect((await pub(app, "0123456789", {}, undefined, as(amy.email))).status).toBe(201) // 10
-    expect((await pub(app, "ABCDE", {}, undefined, as(amy.email))).status).toBe(413) // +5 > 12
-    // Ben's workspace meter is independent — before the per-org fix this 413'd
-    // because the meter summed every workspace's bytes against one global cap.
-    expect((await pub(app, "0123456789", {}, undefined, as(ben.email))).status).toBe(201)
-  })
-
-  it("an artifact in one org with 'org' visibility is 404 to a member of another org", async () => {
-    const { app } = quotaApp("mt-isolation", { multiWorkspace: true }, [amy, ben])
-    await app.request("/v1/me", { headers: as(amy.email) })
-    await app.request("/v1/me", { headers: as(ben.email) })
-    const a = await (
-      await pub(app, "amy's private doc", { visibility: "org" }, undefined, as(amy.email))
-    ).json()
-    // Amy sees her own artifact; Ben (different workspace) cannot, by id.
-    expect(
-      (await app.request(`/v1/artifacts/${a.short_id}`, { headers: as(amy.email) })).status,
-    ).toBe(200)
-    expect(
-      (await app.request(`/v1/artifacts/${a.short_id}`, { headers: as(ben.email) })).status,
-    ).toBe(404)
-    expect(
-      (await app.request(`/v1/artifacts/${a.short_id}/content`, { headers: as(ben.email) })).status,
-    ).toBe(404)
-  })
-})
-
-describe("SSRF guard: integer/hex/octal IP encodings can't bypass the private-range check", () => {
-  it("blocks loopback / metadata / private in every encoding", () => {
-    for (const u of [
-      "http://127.0.0.1/", // dotted
-      "http://2130706433/", // decimal int = 127.0.0.1
-      "http://0x7f000001/", // hex = 127.0.0.1
-      "http://0177.0.0.1/", // octal octet = 127.0.0.1
-      "http://127.1/", // short form = 127.0.0.1
-      "http://[::ffff:127.0.0.1]/", // IPv4-mapped IPv6 loopback
-      "http://169.254.169.254/latest/meta-data/", // cloud metadata
-      "http://localhost/", // name
-      "http://10.0.0.5/",
-      "http://192.168.1.1/",
-      "http://[::1]/", // IPv6 loopback
-      "ftp://example.com/", // wrong scheme
-    ]) {
-      expect(isPublicHttpUrl(u)).toBe(false)
-    }
-  })
-
-  it("allows genuinely public hosts", () => {
-    for (const u of [
-      "https://hooks.example.com/abc",
-      "http://8.8.8.8/",
-      "http://1.1.1.1/",
-      "https://203.0.113.10/webhook",
-    ]) {
-      expect(isPublicHttpUrl(u)).toBe(true)
-    }
-  })
-})
